Move getting data from drop event to helpers.
