+++ assets/src/stories-editor/components/block-navigation/item.js
@@ -17,32 +17,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import { BlockPreviewLabel } from '../';
-
-/**
- * Parses drag & drop events to ensure the event contains valid transfer data.
- *
- * @param {Object} event
- * @return {Object} Parsed event data.
- */
-const parseDropEvent = ( event ) => {
-	let result = {
-		srcClientId: null,
-		srcIndex: null,
-		type: null,
-	};
-
-	if ( ! event.dataTransfer ) {
-		return result;
-	}
-
-	try {
-		result = Object.assign( result, JSON.parse( event.dataTransfer.getData( 'text' ) ) );
-	} catch ( err ) {
-		return result;
-	}
-
-	return result;
-};
+import { parseDropEvent } from '../../helpers';
 
 class BlockNavigationItem extends Component {
 	constructor( ...args ) {
