+++ assets/src/stories-editor/components/editor-carousel/indicator.js
@@ -12,30 +12,9 @@ import { useState } from '@wordpress/element';
 import { useDispatch, useSelect } from '@wordpress/data';
 
 /**
- * Parses drag & drop events to ensure the event contains valid transfer data.
- *
- * @param {Object} event
- * @return {Object} Parsed event data.
+ * Internal dependencies
  */
-const parseDropEvent = ( event ) => {
-	let result = {
-		srcClientId: null,
-		srcIndex: null,
-		type: null,
-	};
-
-	if ( ! event.dataTransfer ) {
-		return result;
-	}
-
-	try {
-		result = Object.assign( result, JSON.parse( event.dataTransfer.getData( 'text' ) ) );
-	} catch ( err ) {
-		return result;
-	}
-
-	return result;
-};
+import { parseDropEvent } from '../../helpers';
 
 /**
  * Carousel indicator component.
