+++ assets/src/stories-editor/components/reorderer/item.js
@@ -15,33 +15,7 @@ import { withDispatch, withSelect } from '@wordpress/data';
  * Internal dependencies
  */
 import { BlockPreview } from '../';
-
-/**
- * Parses drag & drop events to ensure the event contains valid transfer data.
- *
- * @param {Object} event
- * @return {Object} Parsed event data.
- */
-const parseDropEvent = ( event ) => {
-	let result = {
-		srcRootClientId: null,
-		srcClientId: null,
-		srcIndex: null,
-		type: null,
-	};
-
-	if ( ! event.dataTransfer ) {
-		return result;
-	}
-
-	try {
-		result = Object.assign( result, JSON.parse( event.dataTransfer.getData( 'text' ) ) );
-	} catch ( err ) {
-		return result;
-	}
-
-	return result;
-};
+import { parseDropEvent } from '../../helpers';
 
 /**
  * A single item within the list of pages to be reordered.
