+++ assets/src/stories-editor/helpers/index.js
@@ -2143,3 +2143,29 @@ export const isPageBlock = ( clientId ) => {
 	const block = getBlock( clientId );
 	return block && 'amp/amp-story-page' === block.name;
 };
+
+/**
+ * Parses drag & drop events to ensure the event contains valid transfer data.
+ *
+ * @param {Object} event
+ * @return {Object} Parsed event data.
+ */
+export const parseDropEvent = ( event ) => {
+	let result = {
+		srcClientId: null,
+		srcIndex: null,
+		type: null,
+	};
+
+	if ( ! event.dataTransfer ) {
+		return result;
+	}
+
+	try {
+		result = Object.assign( result, JSON.parse( event.dataTransfer.getData( 'text' ) ) );
+	} catch ( err ) {
+		return result;
+	}
+
+	return result;
+};
